+++ src/main/java/ws/beauty/salon/repository/AppointmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ws.beauty.salon.model.Appointment;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Long>{
+public interface AppointmentRepository extends JpaRepository<Appointment, Integer>{
     
 }
